# internal/server/handler.go
package server

import (
	"bytes"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"net/http"
	"strconv"
	"time"

	"github.com/go-chi/chi/v5"
	"go.uber.org/zap"

	"github.com/mikeziminio/go-custom-metrics/internal/ip"
	"github.com/mikeziminio/go-custom-metrics/internal/model"
)

// Update handles HTTP POST /value request to update a single metric.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request containing metric data in JSON body
//
// It reads the metric from request body, updates it in storage,
// and logs an audit event if enabled.
//
//	@Tags		Metrics
//	@Summary	Обновление метрики
//	@ID			updateMetric
//	@Param		metric	body		updateReqSchema	true	"Данные метрики"
//	@Success	200		{object}	model.Metric	"Метрика обновлена"
//	@Failure	400		{string}	string			"Некорректный тип метрики или неверный формат данных"
//	@Failure	500		{string}	string			"Внутренняя ошибка сервера"
//	@Router		/update [post]
func (a *APIServer) Update(res http.ResponseWriter, req *http.Request) {
	body, err := io.ReadAll(req.Body)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to read request body: %w", err))
		return
	}

	var data updateReqSchema

	err = json.Unmarshal(body, &data)
	if err != nil {
		http.Error(res, fmt.Sprintf("failed to validate request body: %v", err),
			http.StatusBadRequest)
		return
	}

	m, err := a.storage.Update(req.Context(), model.Metric{
		ID:    data.ID,
		MType: data.MType,
		Delta: data.Delta,
		Value: data.Value,
	})
	if err != nil {
		if errors.Is(err, model.ErrIncorrectMetricType) {
			http.Error(res, fmt.Sprintf("failed to fetch metric type: %v", data.MType),
				http.StatusBadRequest,
			)
			return
		}
		a.handleInternalServerError(res,
			fmt.Errorf("failed to update metric value %s / %s: %w", data.MType, data.ID, err),
		)
		return
	}

	// Log audit event
	if a.auditLogger != nil {
		ipAddress := ip.ExtractIPAddress(req)
		event := AuditEvent{
			Timestamp: time.Now().Unix(),
			Metrics:   []string{fmt.Sprintf("%s:%s", data.MType, data.ID)},
			IPAddress: ipAddress,
		}
		if err := a.auditLogger.Log(req.Context(), event); err != nil {
			a.logger.Error("Failed to log audit event", zap.Error(err))
		}
	}

	resData, err := json.Marshal(m)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to marshal response: %w", err))
		return
	}

	res.Header().Set("Content-Type", "application/json")
	_, err = res.Write(resData)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to write response: %w", err))
		return
	}
}

// Updates handles HTTP POST /updates request to update multiple metrics.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request containing metrics in JSON body
//
// It reads metrics from request body and updates them in storage in bulk,
// then logs an audit event if enabled.
//
//	@Tags		Metrics
//	@Summary	Массовое обновление метрик
//	@ID			updateMetrics
//	@Param		metrics	body		updatesReqSchema	true	"Список метрик"
//	@Success	200		{string}	string				"OK"
//	@Failure	400		{string}	string				"Некорректный тип метрики или неверный формат данных"
//	@Failure	500		{string}	string				"Внутренняя ошибка сервера"
//	@Router		/updates [post]
func (a *APIServer) Updates(res http.ResponseWriter, req *http.Request) {
	body, err := io.ReadAll(req.Body)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to read request body: %w", err))
		return
	}

	var data updatesReqSchema
	err = json.Unmarshal(body, &data)
	if err != nil {
		http.Error(res, fmt.Sprintf("failed to validate request body: %v", err),
			http.StatusBadRequest)
		return
	}

	metrics := make([]model.Metric, 0, len(data))
	for _, d := range data {
		metrics = append(metrics, model.Metric{
			ID:    d.ID,
			MType: d.MType,
			Delta: d.Delta,
			Value: d.Value,
		})
	}

	err = a.storage.Updates(req.Context(), metrics)

	if err != nil {
		if errors.Is(err, model.ErrIncorrectMetricType) {
			http.Error(res, "failed to fetch metric type",
				http.StatusBadRequest,
			)
			return
		}
		a.handleInternalServerError(res,
			fmt.Errorf("failed to update metric values: %w", err),
		)
		return
	}

	// Log audit event
	if a.auditLogger != nil {
		ipAddress := ip.ExtractIPAddress(req)
		metricNames := make([]string, 0, len(data))
		for _, d := range data {
			metricNames = append(metricNames, fmt.Sprintf("%s:%s", d.MType, d.ID))
		}
		event := AuditEvent{
			Timestamp: time.Now().Unix(),
			Metrics:   metricNames,
			IPAddress: ipAddress,
		}
		if err := a.auditLogger.Log(req.Context(), event); err != nil {
			a.logger.Error("Failed to log audit event", zap.Error(err))
		}
	}

	res.WriteHeader(http.StatusOK)
}

// UpdateByParams handles HTTP POST /update/{metricType}/{metricName}/{value} request.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request with metric parameters in URL path
//
// It extracts metric parameters from URL and updates the metric in storage,
// then logs an audit event if enabled.
//
//	@Tags		Metrics
//	@Summary	Обновление метрики по параметрам URL
//	@ID			updateMetricByParams
//	@Param		metricType	path		string	true	"Тип метрики (counter или gauge)"
//	@Param		metricName	path		string	true	"Имя метрики"
//	@Param		value		path		string	true	"Значение метрики"
//	@Success	200			{string}	string	"OK"
//	@Failure	400			{string}	string	"Некорректный тип метрики или неверный формат значения"
//	@Failure	500			{string}	string	"Внутренняя ошибка сервера"
//	@Router		/update/{metricType}/{metricName}/{value} [post]
func (a *APIServer) UpdateByParams(res http.ResponseWriter, req *http.Request) {
	mt := chi.URLParam(req, "metricType")
	metricType, err := model.NewMetricTypeFromString(mt)
	if err != nil {
		http.Error(res, fmt.Sprintf("failed to fetch metric type: %v", err),
			http.StatusBadRequest,
		)
		return
	}
	metricName := chi.URLParam(req, "metricName")
	var delta *int64
	var value *float64

	val := chi.URLParam(req, "value")
	if metricType == model.Counter {
		d, err := strconv.ParseInt(val, 10, 64)
		if err != nil {
			http.Error(res, fmt.Sprintf("failed to parse counter value: %v", err),
				http.StatusBadRequest,
			)
			return
		}
		delta = &d
	}

	if metricType == model.Gauge {
		v, err := strconv.ParseFloat(val, 64)
		if err != nil {
			http.Error(res, fmt.Sprintf("failed to parse gauge value: %v", err),
				http.StatusBadRequest,
			)
			return
		}
		value = &v
	}

	_, err = a.storage.Update(req.Context(), model.Metric{
		ID:    metricName,
		MType: metricType,
		Delta: delta,
		Value: value,
	})
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to update metric value: %w", err))
		return
	}

	// Log audit event
	if a.auditLogger != nil {
		ipAddress := ip.ExtractIPAddress(req)
		event := AuditEvent{
			Timestamp: time.Now().Unix(),
			Metrics:   []string{fmt.Sprintf("%s:%s", metricType, metricName)},
			IPAddress: ipAddress,
		}
		if err := a.auditLogger.Log(req.Context(), event); err != nil {
			a.logger.Error("Failed to log audit event", zap.Error(err))
		}
	}

	res.WriteHeader(http.StatusOK)
}

// Get handles HTTP POST /value request to retrieve a single metric.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request containing metric type and name in JSON body
//
// It reads metric type and name from request body and returns the value.
//
//	@Tags		Metrics
//	@Summary	Получение метрики
//	@ID			getMetric
//	@Param		metric	body		getReqSchema	true	"Идентификатор метрики"
//	@Success	200		{object}	model.Metric	"Метрика найдена"
//	@Failure	404		{string}	string			"Метрика не найдена"
//	@Failure	500		{string}	string			"Внутренняя ошибка сервера"
//	@Router		/value [post]
func (a *APIServer) Get(res http.ResponseWriter, req *http.Request) {
	body, err := io.ReadAll(req.Body)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to read request body: %w", err))
		return
	}

	var data getReqSchema

	err = json.Unmarshal(body, &data)
	if err != nil {
		http.Error(res, fmt.Sprintf("failed to validate request body: %v", err),
			http.StatusBadRequest)
		return
	}

	m, err := a.storage.Get(req.Context(), data.MType, data.ID)
	if err != nil {
		if errors.Is(err, model.ErrMetricNotFound) {
			http.Error(res, fmt.Sprintf("metric not found %s / %s: %v", data.MType, data.ID, err),
				http.StatusNotFound,
			)
			return
		}
		a.handleInternalServerError(res, fmt.Errorf("failed to get metric: %w", err))
		return
	}

	resData, err := json.Marshal(m)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to marshal response: %w", err))
		return
	}

	res.Header().Set("Content-Type", "application/json")
	_, err = res.Write(resData)
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to write response: %w", err))
		return
	}
}

// GetByParams handles HTTP GET /value/{metricType}/{metricName} request.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request with metric parameters in URL path
//
// It extracts metric parameters from URL and returns the value as plain text.
//
//	@Tags		Metrics
//	@Summary	Получение метрики по параметрам URL
//	@ID			getMetricByParams
//	@Param		metricType	path		string	true	"Тип метрики (counter или gauge)"
//	@Param		metricName	path		string	true	"Имя метрики"
//	@Success	200			{string}	string	"Значение метрики"
//	@Failure	404			{string}	string	"Метрика не найдена"
//	@Failure	400			{string}	string	"Некорректный тип метрики"
//	@Failure	500			{string}	string	"Внутренняя ошибка сервера"
//	@Router		/value/{metricType}/{metricName} [get]
func (a *APIServer) GetByParams(res http.ResponseWriter, req *http.Request) {
	mt := chi.URLParam(req, "metricType")
	metricType, err := model.NewMetricTypeFromString(mt)
	if err != nil {
		http.Error(res, fmt.Sprintf("failed to fetch metric type: %v", err),
			http.StatusBadRequest,
		)
		return
	}
	metricName := chi.URLParam(req, "metricName")

	m, err := a.storage.Get(req.Context(), metricType, metricName)
	if err != nil {
		if errors.Is(err, model.ErrMetricNotFound) {
			http.Error(res, fmt.Sprintf("metric not found %s / %s: %v", metricType, metricName, err),
				http.StatusNotFound,
			)
			return
		}
		a.handleInternalServerError(res, fmt.Errorf("failed to get metric %s / %s: %w", metricType, metricName, err))
		return
	}

	var r string
	if m.MType == model.Gauge {
		r = strconv.FormatFloat(*m.Value, 'f', -1, 64)
	} else {
		r = fmt.Sprintf("%d", *m.Delta)
	}

	res.Header().Set("Content-Type", "text/html")
	_, err = res.Write([]byte(r))
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to write response: %w", err))
		return
	}
}

// List handles HTTP GET / request to list all metrics.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request
//
// It returns metrics in plain text format with id and value.
//
//	@Tags		Metrics
//	@Summary	Список всех метрик
//	@ID			listMetrics
//	@Success	200	{string}	string	"Список метрик в HTML формате"
//	@Failure	500	{string}	string	"Внутренняя ошибка сервера"
//	@Router		/ [get]
func (a *APIServer) List(res http.ResponseWriter, req *http.Request) {
	var b bytes.Buffer
	metrics, err := a.storage.List(req.Context())
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to get metrics list: %w", err))
		return
	}
	for id, m := range metrics {
		b.WriteString(id)
		b.WriteString(" ")
		switch m.MType {
		case model.Gauge:
			b.WriteString(strconv.FormatFloat(*m.Value, 'f', 5, 64))
		case model.Counter:
			b.WriteString(strconv.FormatInt(*m.Delta, 10))
		}
	}

	res.Header().Set("Content-Type", "text/html")
	_, err = res.Write(b.Bytes())
	if err != nil {
		a.handleInternalServerError(res, fmt.Errorf("failed to write response: %w", err))
		return
	}
}

// Ping handles HTTP GET /ping health check request.
//
// Parameters:
//   - res: HTTP response writer
//   - req: HTTP request
//
// It returns status 200 if storage is reachable.
//
//	@Tags		Info
//	@Summary	Запрос состояния сервиса
//	@ID			infoPing
//	@Success	200	{string}	string	"OK"
//	@Failure	500	{string}	string	"Внутренняя ошибка"
//	@Router		/ping [get]
func (a *APIServer) Ping(res http.ResponseWriter, req *http.Request) {
	err := a.storage.Ping(req.Context())
	if err != nil {
		a.handleInternalServerError(res, err)
		return
	}
	res.WriteHeader(http.StatusOK)
}

func (a *APIServer) handleInternalServerError(res http.ResponseWriter, err error) {
	a.logger.Error("Internal server error", zap.Error(err))
	status := http.StatusInternalServerError
	http.Error(res, http.StatusText(status), status)
}